yet anather refactoring
